\name{dic13pof2008.2009}
\docType{data}

\alias{dic13pof2008.2009}

\encoding{UTF-8}
\title{Dictionary for vehicle outgoing record POF 2008-2009 survey}

\description{

  Data dictionary for reading "T_DESPESA_VEICULO_S.txt" file (vehicle
  outgoing (type 13) record from POF 2008-2009 survey).

}

\format{
  This dictionary is a data frame with one line for each variable,
  and it has 4 columms, with important information of each variable.
  
  \tabular{lll}{
    inicio \tab numeric \tab Position of the value's first character in
    the data file \cr
    cod \tab character \tab Variable code \cr
    tamanho \tab numeric \tab Field size \cr
    desc \tab character \tab Short description\cr
  
  }
}
\source{
  http://www.ibge.gov.br
    }
\keyword{datasets}