\name{dic9pof2008.2009}
\docType{data}

\alias{dic9pof2008.2009}

\encoding{UTF-8}
\title{Dictionary for  outgoing with domestic services record POF 2008-2009 survey}

\description{

  Data dictionary for reading "T_SERVICO_DOMS_S.txt" file (outgoing with
  domestic services (type 9) record from POF 2008-2009 survey).

}

\format{
  This dictionary is a data frame with one line for each variable,
  and it has 4 columms, with important information of each variable.
  
  \tabular{lll}{
    inicio \tab numeric \tab Position of the value's first character in
    the data file \cr
    cod \tab character \tab Variable code \cr
    tamanho \tab numeric \tab Field size \cr
    desc \tab character \tab Short description\cr
  
  }
}
\source{The dictionaries were made based on the documentation of the
    PNAD survey available at \url{http://www.ibge.gov.br}.}
\keyword{datasets}