\name{dicPME}
\docType{data}
\alias{dicPME}

\encoding{UTF-8}
\title{Dicionários de Dados da PME}
\description{
  Um dicionário de dados para as pesquisas PME de 2002 a 2009.
}
\format{
  O dicionário é um data frame onde cada linha corresponde a uma
  variável, e que tem as seguintes colunas:
  \tabular{lll}{
    inicio \tab numeric \tab Posição do primeiro caracter
    do campo \cr
    cod \tab character \tab Código da variável \cr
    tamanho \tab numeric \tab Tamanho do campo \cr
    desc \tab character \tab Descrição resumida do campo
  }
}
\source{Documentação da PME fornecida no CD de dados do IBGE}
\keyword{datasets}