\name{dicPNAD2004}
\alias{dicPNAD2004}
\docType{data}
\encoding{UTF-8}

\title{Data set for PNAD 2004 suvey}

\description{
  This dataset contains 2 data.frames.

  The two data.frames, \link{dic_dom_2004} and \link{dic_pessoas_2004}, are data
  dictionaries used by the fuction \link{le.pesquisa} for reading the
  two microdada files provided by IBGE.
  
}

\format{A data set with 2 data.frames:

  \enumerate{
    \item dic_dom_2004
    \item dic_pessoas_2004
  }
}

\source{The dictionaries were made based on the documentation of the
    PNAD survey available at \url{http://www.ibge.gov.br}.}

\keyword{datasets}


\keyword{datasets}
