\name{dicPNAD2008}
\docType{data}

\alias{dicPNAD2008}

\encoding{UTF-8}
\title{Data set for PNAD 2008 survey}
\description{
  One data set with two dictionaries one for people records and one for domiciles records.
}
\format{
  A data set with 2 different data frames:

  -dicdom2008 (dictionary with domiciles record from PNAD 2008 survey)

  -dicpes2008 (dictionary with people record from PNAD 2008 survey)

}
\source{
    http://www.ibge.gov.br
  
    }
\keyword{datasets}