\name{dicPOF}
\docType{data}
\alias{dicPOF}
\alias{dicPOF1987.1988}
\alias{dicPOF1995.1996}
\alias{dicPOF2002.2003}
\alias{dicPOF2008.2009}

\alias{dic1pof1987.1988}
\alias{dic2pof1987.1988}
\alias{dic3pof1987.1988}
\alias{dic4pof1987.1988}
\alias{dic5pof1987.1988}
\alias{dic1pof1995.1996}
\alias{dic2pof1995.1996}
\alias{dic3pof1995.1996}
\alias{dic4pof1995.1996}
\alias{dic5pof1995.1996}
\alias{dic6pof1995.1996}
\alias{dic7pof1995.1996}
\alias{dic8pof1995.1996}
\alias{dic9pof1995.1996}
\alias{dic10pof1995.1996}
\alias{dic11pof1995.1996}
\alias{dic12pof1995.1996}
\alias{dic1pof2002.2003}
\alias{dic2pof2002.2003}
\alias{dic3pof2002.2003}
\alias{dic4pof2002.2003}
\alias{dic5pof2002.2003}
\alias{dic6pof2002.2003}
\alias{dic7pof2002.2003}
\alias{dic8pof2002.2003}
\alias{dic9pof2002.2003}
\alias{dic10pof2002.2003}
\alias{dic11pof2002.2003}
\alias{dic12pof2002.2003}
\alias{dic13pof2002.2003}
\alias{dic14pof2002.2003}
\alias{dic1pof2008.2009}
\alias{dic2pof2008.2009}
\alias{dic3pof2008.2009}
\alias{dic4pof2008.2009}
\alias{dic5pof2008.2009}
\alias{dic6pof2008.2009}
\alias{dic7pof2008.2009}
\alias{dic8pof2008.2009}
\alias{dic9pof2008.2009}
\alias{dic10pof2008.2009}
\alias{dic11pof2008.2009}
\alias{dic12pof2008.2009}
\alias{dic13pof2008.2009}
\alias{dic14pof2008.2009}
\alias{dic15pof2008.2009}

\encoding{UTF-8}
\title{Data Dictionaries for POF}
\description{
  One data dictionary for each record type from POF surveys of
  1987/1988, 1995/1996 and 2002/2003.
}
\format{
  Each dictionary is a data frame with one line for each survey variable,
  and it has the following columns:  
  
  \tabular{lll}{
    inicio \tab numeric \tab Position of the value's first character in
    the data file \cr
    cod \tab character \tab Variable code \cr
    tamanho \tab numeric \tab Field size \cr
    desc \tab character \tab Short description
  }
}
\source{POF documentation available in the CD made by IBGE}
\keyword{datasets}