\name{rotpes2009}
\docType{data}

\alias{rotpes2009}

\encoding{UTF-8}
\title{Data Dictionarie label for PNAD 2009 survey}
\description{
  One data dictionary label for people record from PNAD 2009 survey.
}
\format{
  This dictionary is a data frame with one line for each survey categorical variable,
  and it has 3 columms, with important information of each categorical variable.
  
  \tabular{lll}{
    cod \tab character \tab Variable code \cr
    valor \tab numeric \tab Levels of the categorical variable \cr
    rotulo \tab numeric \tab Labels of the categorical variable 
  }
}
\source{
  http://www.ibge.gov.br
    }
\keyword{datasets}