\name{dicPNAD2006}
\docType{data}

\alias{dicPNAD2006}

\encoding{UTF-8}
\title{Data set for PNAD 2006 survey}
\description{
  One data set with four dictionaries two for people records (one of
  them has the reconsidered survey) and two for domiciles records(one of
  them has the reconsidered survey). If you want to read all the
  variables used in the survey use the \code{dicdom2006supl} or
  \code{dicpes2006supl} dictionary.
}
\format{
  A data set with 4 different data frames:

  -dicdom2006 (dictionary with domiciles record from PNAD 2006 survey)

  -dicpes2006 (dictionary with people record from PNAD 2006 survey)
  
  -dicdom2006supl (dictionary with domiciles record from PNAD 2006 (reconsidered) survey)

  -dicpes2006supl (dictionary with people record from PNAD 2006 (reconsidered) survey)

}
\source{
  PNAD documentation available in the CD's made by IBGE
  
    }
\keyword{datasets}