\name{dicPNAD2007}
\docType{data}

\alias{dicPNAD2007}

\encoding{UTF-8}
\title{Data set for PNAD 2007 survey}
\description{
  One data set with two dictionaries one for people records and one for domiciles records.
}
\format{
  A data set with 2 different data frames:

  -dicdom2007 (dictionary with domiciles record from PNAD 2007 survey)

  -dicpes2007 (dictionary with people record from PNAD 2007 survey)

}
\source{
    http://www.ibge.gov.br
  
    }
\keyword{datasets}