\name{dicPNAD2007}
\docType{data}

\alias{dicPNAD2007}

\encoding{UTF-8}
\title{Data set for PNAD 2007 survey}
\description{
  
  This dataset contains 2 data.frames.

  The two data.frames, \link{dicdom2007} and \link{dicpes2007}, are data
  dictionaries used by the fuction \link{le.pesquisa} for reading the
  two microdada files provided by IBGE.

}
\format{
  A data set with 2 different data frames:
  \enumerate{
  \item dicdom2007 (dictionary with domiciles record from PNAD 2007 survey)

  \item dicpes2007 (dictionary with people record from PNAD 2007 survey)
  }
}
\source{The dictionaries were made based on the documentation of the
    PNAD survey available at \url{http://www.ibge.gov.br}.}
\keyword{datasets}