\name{dicPNAD2008}
\docType{data}

\alias{dicPNAD2008}

\encoding{UTF-8}
\title{Data set for PNAD 2008 survey}
\description{
  This dataset contains 2 data.frames.

  The two data.frames, \link{dicdom2008} and \link{dicpes2008}, are data
  dictionaries used by the fuction \link{le.pesquisa} for reading the
  two microdada files provided by IBGE.

}
\format{
  A data set with 2 different data frames:
  \enumerate{
    
  \item dicdom2008 (dictionary with domiciles record from PNAD 2008 survey)

  \item dicpes2008 (dictionary with people record from PNAD 2008 survey)
  }
}
\source{The dictionaries were made based on the documentation of the
    PNAD survey available at \url{http://www.ibge.gov.br}.}
\keyword{datasets}