\name{dicPNAD2009}
\docType{data}

\alias{dicPNAD2009}

\encoding{UTF-8}
\title{Data set for PNAD 2009 survey}
\description{
  One data set with two dictionaries one for people records and one for house records, and
  two dictionaries labels one for people records and one for house records.
}
\format{
  A data set with 4 different data frames:

  -dicdom2009 (dictionary with domiciles record from PNAD 2009 survey)

  -dicpes2009 (dictionary with people record from PNAD 2009 survey)

  -rotdom2009 (data dictionary label with domiciles record from PNAD
  2009 survey)

  -rotpes2009 (data dictionary label with people record from PNAD 2009
  survey)
}
\source{
  http://www.ibge.gov.br
  
    }
\keyword{datasets}