\name{dicPNAD2009}
\docType{data}

\alias{dicPNAD2009}

\encoding{UTF-8}
\title{Data set for PNAD 2009 suvey}
\description{
  One data set with two dictionaries one for people records and one for house records, and
  two dictionaries labels one for people records and one for house records.
}
\format{
  A data set with 4 different data frames:

  -dicdom2009

  -dicpes2009

  -rotdom2009

  -rotpes2009
}
\source{
  http://www.ibge.gov.br
  
    }
\keyword{datasets}