\name{dicPNAD2009}
\docType{data}

\alias{dicPNAD2009}

\encoding{UTF-8}
\title{Data set for PNAD 2009 suvey}

\description{
  This dataset contains 4 data.frames.

  The two data.frames, \link{dicdom2009} and \link{dicpes2009}, are data
  dictionaries used by the fuction \link{le.pesquisa} for reading the
  two microdada files provided by IBGE.

  The two data.frames, \link{rotdom2009} and \link{rotpes2009}, are
  labels of some data values. They are needed only if one wants to work
  with factors.
}

\format{ A data set with 4 different data frames:
  
  \enumerate{
    \item dicdom2009
    \item dicpes2009
    \item rotdom2009
    \item rotpes2009
  }
}

\source{The dictionaries were made based on the documentation of the
    PNAD survey available at \url{http://www.ibge.gov.br}.}

\keyword{datasets}
  