\name{dicPNAD2012}
\alias{dicPNAD2012}
\docType{data}
\encoding{UTF-8}

\title{Data set for PNAD 2012 suvey}

\description{
  This dataset contains 2 data.frames.

  The two data.frames, \link{dicdom2012} and \link{dicpes2012}, are data
  dictionaries used by the fuction \link{le.pesquisa} for reading the
  two microdada files provided by IBGE.
  
}

\format{A data set with 2 data.frames:

  \enumerate{
    \item dicdom2012
    \item dicpes2012
  }
}

\source{The dictionaries were made based on the documentation of the
    PNAD survey available at \url{http://www.ibge.gov.br}.}

\keyword{datasets}

