\name{dicPOF}
\docType{data}
\alias{dicPOF}
\alias{dicPOF1987.1988}
\alias{dicPOF1995.1996}
\alias{dicPOF2002.2003}

\alias{dic1pof1987.1988}
\alias{dic2pof1987.1988}
\alias{dic3pof1987.1988}
\alias{dic4pof1987.1988}
\alias{dic5pof1987.1988}
\alias{dic1pof1995.1996}
\alias{dic2pof1995.1996}
\alias{dic3pof1995.1996}
\alias{dic4pof1995.1996}
\alias{dic5pof1995.1996}
\alias{dic6pof1995.1996}
\alias{dic7pof1995.1996}
\alias{dic8pof1995.1996}
\alias{dic9pof1995.1996}
\alias{dic10pof1995.1996}
\alias{dic11pof1995.1996}
\alias{dic12pof1995.1996}
\alias{dic1pof2002.2003}
\alias{dic2pof2002.2003}
\alias{dic3pof2002.2003}
\alias{dic4pof2002.2003}
\alias{dic5pof2002.2003}
\alias{dic6pof2002.2003}
\alias{dic7pof2002.2003}
\alias{dic8pof2002.2003}
\alias{dic9pof2002.2003}
\alias{dic10pof2002.2003}
\alias{dic11pof2002.2003}
\alias{dic12pof2002.2003}
\alias{dic13pof2002.2003}
\alias{dic14pof2002.2003}

\encoding{UTF-8}
\title{Dicionários de Dados da POF}
\description{
  Um dicionário de dados para cada tipo de registro das POF's de 1987/1988, 1995/1996 e 2002/2003.
}
\format{
  Cada dicionário é um data frame onde cada linha corresponde a uma
  variável, e que tem as seguintes colunas:
  \tabular{lll}{
    inicio \tab numeric \tab Posição do primeiro caracter
    do campo \cr
    cod \tab character \tab Código da variável \cr
    tamanho \tab numeric \tab Tamanho do campo \cr
    desc \tab character \tab Descrição resumida do campo
  }
}
\source{Documentação da POF fornecida no CD de dados do IBGE}
\keyword{datasets}